+++ SecretSauce/src/main/java/com/byoutline/secretsauce/activities/WebViewActivityV7.java
@@ -2,7 +2,7 @@ package com.byoutline.secretsauce.activities;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.KeyEvent;
 import android.view.MenuItem;