+++ SecretSauce/src/main/java/com/byoutline/secretsauce/di/AppInjector.kt
@@ -19,9 +19,9 @@ package com.byoutline.secretsauce.di
 import android.app.Activity
 import android.app.Application
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentActivity
-import android.support.v4.app.FragmentManager
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.FragmentManager
 
 import dagger.android.AndroidInjection
 import dagger.android.support.AndroidSupportInjection
@@ -59,10 +59,10 @@ object AppInjector {
         if (activity is HasSupportFragmentInjector || activity is Injectable) {
             AndroidInjection.inject(activity)
         }
-        (activity as? FragmentActivity)?.supportFragmentManager?.registerFragmentLifecycleCallbacks(
-            object : FragmentManager.FragmentLifecycleCallbacks() {
+        (activity as? androidx.fragment.app.FragmentActivity)?.supportFragmentManager?.registerFragmentLifecycleCallbacks(
+            object : androidx.fragment.app.FragmentManager.FragmentLifecycleCallbacks() {
                 override fun onFragmentCreated(
-                    fm: FragmentManager, f: Fragment,
+                    fm: androidx.fragment.app.FragmentManager, f: androidx.fragment.app.Fragment,
                     savedInstanceState: Bundle?
                 ) {
                     if (f is Injectable) {