+++ SecretSauce/src/main/java/com/byoutline/secretsauce/di/ViewModelKey.java
@@ -1,6 +1,6 @@
 package com.byoutline.secretsauce.di;
 
-import android.arch.lifecycle.ViewModel;
+import androidx.lifecycle.ViewModel;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;