+++ SecretSauce/src/main/java/com/byoutline/secretsauce/lifecycle/RxLifecycle.kt
@@ -3,11 +3,11 @@ package com.byoutline.secretsauce.lifecycle
 import android.app.Activity
 import android.app.Application
 import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentActivity
-import android.support.v4.app.FragmentManager
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentActivity
+import androidx.fragment.app.FragmentManager
 
-fun FragmentActivity.invokeOnStopped(action: () -> Unit) {
+fun androidx.fragment.app.FragmentActivity.invokeOnStopped(action: () -> Unit) {
     val activity = this
     application.registerActivityLifecycleCallbacks(object : Application.ActivityLifecycleCallbacks {
         override fun onActivityCreated(a: Activity?, savedInstanceState: Bundle?) {}
@@ -27,15 +27,15 @@ fun FragmentActivity.invokeOnStopped(action: () -> Unit) {
     })
 }
 
-fun Fragment.invokeOnStopped(action: () -> Unit) {
+fun androidx.fragment.app.Fragment.invokeOnStopped(action: () -> Unit) {
     val fragment = this
     activity!!.supportFragmentManager.registerFragmentLifecycleCallbacks(object :
-        FragmentManager.FragmentLifecycleCallbacks() {
-        override fun onFragmentStopped(fm: FragmentManager, f: Fragment) {
+        androidx.fragment.app.FragmentManager.FragmentLifecycleCallbacks() {
+        override fun onFragmentStopped(fm: androidx.fragment.app.FragmentManager, f: androidx.fragment.app.Fragment) {
             if (f === fragment) action()
         }
 
-        override fun onFragmentDestroyed(fm: FragmentManager, f: Fragment) {
+        override fun onFragmentDestroyed(fm: androidx.fragment.app.FragmentManager, f: androidx.fragment.app.Fragment) {
             if (f === fragment) fm.unregisterFragmentLifecycleCallbacks(this)
         }
     }, false)