+++ SecretSauce/src/main/java/com/byoutline/secretsauce/listeners/PicassoScrollListener.java
@@ -1,7 +1,7 @@
 package com.byoutline.secretsauce.listeners;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.squareup.picasso.Picasso;