+++ SecretSauce/src/main/java/com/byoutline/secretsauce/utils/CustomTypefaceSpan.java
@@ -2,7 +2,7 @@ package com.byoutline.secretsauce.utils;
 
 import android.graphics.Paint;
 import android.graphics.Typeface;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextPaint;
 import android.text.style.MetricAffectingSpan;