+++ sampleapplication/src/androidTest/java/com/byoutline/sampleapplication/MainActivityShould.kt
@@ -1,6 +1,6 @@
 package com.byoutline.sampleapplication
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import br.com.concretesolutions.kappuccino.custom.recyclerView.RecyclerViewInteractions.recyclerView
 import com.byoutline.espressohelpers.activityTestRule
 import com.schibsted.spain.barista.assertion.BaristaRecyclerViewAssertions.assertRecyclerViewItemCount