+++ sampleapplication/src/main/java/com/byoutline/sampleapplication/MainActivity.kt
@@ -2,7 +2,7 @@ package com.byoutline.sampleapplication
 
 import android.content.Intent
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import com.byoutline.sampleapplication.databinding.ActivityMainBinding
 import com.byoutline.secretsauce.activities.WebViewActivityV7
 import com.byoutline.secretsauce.databinding.bindContentView